pridana kontrola vstupov cez js

+++ App/Views/Reviews/article.view.php
@@ -11,11 +11,11 @@ $review = $data;
         <p class="review_text"><?=$review->getParagraph1()?></p>
         <p class="review_text"><?=$review->getParagraph2()?></p>
         <p class="review_text"><?=$review->getParagraph3()?></p>
-        <img src="<?=$review->getImage()?>" alt="<?=$review->getImagealt()?>" class="review_img">
+        <img src="<?=$review->getImageSrc()?>" alt="<?=$review->getImageAlt()?>" class="review_img">
         <p class="review_text"><?=$review->getParagraph4()?></p>
         <div class="articlebtns">
             <a href="?c=reviews&a=editArticleForm&id=<?=$review->getId()?>" class="btn btn-primary articlebtn">E</a>
-            <a href="?c=reviews&a=deleteArticle&id=<?=$review->getId()?>" class="btn btn-danger articlebtn">X</a>
+            <a href="?c=reviews&a=deleteArticle&id=<?=$review->getId()?>" onclick="return confirm('Recenzia bude zmazaná')" class="btn btn-danger articlebtn">X</a>
         </div>
 
     </article>
